add task2

+++ Task2/Task2_1/hello_world_syscall.c
@@ -7,11 +7,12 @@ enum {
     STDOUT = 1
 };
 
-static int my_write(const int fd, const void* buf, const size_t count) {
-    return syscall(SYS_write, fd, buf, count); // SYS_write = 4
+static ssize_t my_write(const int fd, const void* buf, const size_t count) {
+    return syscall(SYS_write, fd, buf, count);
 }
 
 int main(void) {
-    int ret = my_write(STDOUT, "hello world syscall", 19);
+    ssize_t ret = my_write(STDOUT, "hello world syscall", 19);
     return (ret == ERROR) ? EXIT_FAILURE : EXIT_SUCCESS;
+
 }
